# [exporter/prometheusremotewrite] Fix data race for batch state by removing the state altogether