@dashpole 2024-12-02
cc @ben-childs-docusign 
Can you help us reproduce the performance improvements from https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/34271 in a benchmark?