@ArthurSens 2024-12-13
Closing in favor of https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36601#issuecomment-2542110297